==> notebooks/1_data_cleaning.ipynb <==
# Load the dataset

import pandas as pd
data = pd.read_csv('../data/feedback_prize/feedback_prize_input.csv')
data.head()

data['label'] = data.cat_label.astype('category').cat.codes
data.head()

**Import data to a csv file**

data[['text', 'label']].to_csv('../data/clean/feedback_prize_7_labels.csv')

# Combine claims and premises in 2 labels

def combine_label(row):
    """
    Put claims and counterclaims together under 'claim' and the remaining labels under 'premise'
    """
    if row['cat_label'].lower() == 'claim' or row['cat_label'] == 'counterclaim':
        return 'claim'
    else:
        return 'premise'

data['com_label'] = data.apply(lambda row: combine_label(row), axis=1)
data.head()

data['label'] = data.com_label.astype('category').cat.codes
data.head()

data.drop(['cat_label', 'com_label'], axis=1, inplace=True)
data.head()

count_df = data.groupby('label').count()

count_df['portion'] = count_df['text'] / count_df['text'].sum()
count_df

>=> Our data is mildly imbalanced (size of minority class between 20% and 40%) [link](https://developers.google.com/machine-learning/data-prep/construct/sampling-splitting/imbalanced-data#:~:text=A%20classification%20data%20set%20with,smaller%20proportion%20are%20minority%20classes.)

**Import data to a csv file**

data.to_csv('../data/clean/feedback_prize_2_labels.csv', index=False)
