==> notebooks/3_text_classification_inference.ipynb <==
# Inference

sentences = ['Modern humans today are always on their phone.',
        'They are always on their phone more than 5 hours a day no stop.',
        'All they do is text back and forward and just have group Chats on social media.',
        'They even do it while driving.',
        'They are some really bad consequences when stuff happens when it comes to a phone.']

**Through a pipeline**

from transformers import pipeline

classifier = pipeline("sentiment-analysis", model="fatenghali/text_classification_model")

results = classifier(sentences)

for d, s in zip(results, sentences):
    d['sentence'] = s
    print(d)
    print

**Manually**

from transformers import AutoTokenizer

tokenizer = AutoTokenizer.from_pretrained("fatenghali/text_classification_model")
inputs = []
#inputs = tokenizer(sent, return_tensors="pt")
for sent in sentences:
    inputs.append(tokenizer(sent, return_tensors="pt"))

inputs

from transformers import AutoModelForSequenceClassification
import torch

model = AutoModelForSequenceClassification.from_pretrained("fatenghali/text_classification_model")
with torch.no_grad():
    #logits = model(**inputs).logits #for 1 sentence
    logits = []
    for input_ in inputs:
        logits.append(model(**input_).logits)

#predicted_class_id = logits.argmax().item()
predicted_class_ids = list(map(lambda x: x.argmax().item(), logits))

#model.config.id2label[predicted_class_id]
for _ in predicted_class_ids:
    print(model.config.id2label[_])

**=>** Both methods (through a pipeline and manually) seem to give the same output. But, we still prefer the one through a pipeline for simplicity of code.
